This commit adds: - moves `_create_argument_parser()` to `cli.py` - adds helper methods to process option string destinations and type conversions - adds `from_option_string` methods to *Options classes as required
